Extract downloader logic from Application to a separate class and use it through DI
