+++ app/src/main/java/org/covidwatch/android/CovidWatchApplication.kt
@@ -1,29 +1,20 @@
 package org.covidwatch.android
 
 import android.app.Application
-import androidx.work.*
-import org.covidwatch.android.data.contactevent.ContactEventsDownloadWorker
+import org.covidwatch.android.data.contactevent.ContactEventsDownloader
 import org.covidwatch.android.data.contactevent.LocalContactEventsUploader
 import org.covidwatch.android.di.appModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
-import java.util.concurrent.TimeUnit
+import org.koin.android.ext.android.inject
 
 class CovidWatchApplication : Application() {
 
     private lateinit var localContactEventsUploader: LocalContactEventsUploader
-
-    companion object {
-        private var context: CovidWatchApplication? = null
-
-        fun getContext(): CovidWatchApplication? {
-            return context
-        }
-    }
+    private val contactEventsDownloader: ContactEventsDownloader by inject()
 
     override fun onCreate() {
         super.onCreate()
-        context = this
 
         startKoin {
             androidContext(applicationContext)
@@ -33,38 +24,6 @@ class CovidWatchApplication : Application() {
         localContactEventsUploader = LocalContactEventsUploader(this)
         localContactEventsUploader.startUploading()
 
-        schedulePeriodicPublicContactEventsRefresh()
-    }
-
-    private fun schedulePeriodicPublicContactEventsRefresh() {
-        val downloadRequest =
-            PeriodicWorkRequestBuilder<ContactEventsDownloadWorker>(1, TimeUnit.HOURS)
-                .setConstraints(getRefreshConstraints())
-                .build()
-
-        WorkManager.getInstance(this).enqueueUniquePeriodicWork(
-            "ContactEventsDownloadWorkerPeriodicRefresh",
-            ExistingPeriodicWorkPolicy.REPLACE,
-            downloadRequest
-        )
-    }
-
-    fun executePublicContactEventsRefresh() {
-        val downloadRequest = OneTimeWorkRequestBuilder<ContactEventsDownloadWorker>()
-            .setConstraints(getRefreshConstraints())
-            .build()
-
-        WorkManager.getInstance(this).enqueueUniqueWork(
-            "ContactEventsDownloadWorkerOneTimeRefresh",
-            ExistingWorkPolicy.REPLACE,
-            downloadRequest
-        )
-    }
-
-    private fun getRefreshConstraints(): Constraints {
-        return Constraints.Builder()
-            .setRequiresCharging(false)
-            .setRequiredNetworkType(NetworkType.CONNECTED)
-            .build()
+        contactEventsDownloader.schedulePeriodicPublicContactEventsRefresh()
     }
 }
